requirements.txt
pynetstring
python-miio
